cache

+++ app/Livewire/ArticleList.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Article;
+use Livewire\Attributes\Computed;
 use Livewire\Attributes\Title;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -12,8 +13,21 @@ class ArticleList extends AdminComponent
 {
     use WithPagination;
     public $showOnlyPublished = false;
+    #[Computed]
+    public function articles(){
+        $query = Article::query();
+        if($this->showOnlyPublished){
+            $query->where('published', 1);
+        }
+        return $query->paginate(10, pageName: 'articles-page');
+    }
     public function delete(Article $article){
-        $article->delete();
+        if ($this->articles->count() < 10) {
+            throw new \Exception('Nope');
+        } 
+        $article->delete(); 
+        unset($this->articles); 
+        $this->dispatch('refreshPublishedCount');
     }
     public function showAll(){
         $this->showOnlyPublished = false;
@@ -23,14 +37,5 @@ class ArticleList extends AdminComponent
         $this->showOnlyPublished = true;
         $this->resetPage(pageName: 'articles-page');
     }
-    public function render()
-    {
-        $query = Article::query();
-        if($this->showOnlyPublished){
-            $query->where('published', 1);
-        }
-        return view('livewire.article-list',[
-            'articles' => $query->paginate(10, pageName: 'articles-page')
-        ]);
-    }
+
 }
